LPS-53821 Move to the right package
