+++ modules/apps/document-library/document-library-item-selector-web/src/META-INF/resources/init.jsp
@@ -19,7 +19,7 @@
 <%@ taglib uri="http://liferay.com/tld/item-selector/ui" prefix="item-selector-ui" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
-<%@ page import="com.liferay.document.library.item.selector.web.DLItemSelectorCriterion" %><%@
+<%@ page import="com.liferay.document.library.item.selector.DLItemSelectorCriterion" %><%@
 page import="com.liferay.document.library.item.selector.web.DLItemSelectorView" %><%@
 page import="com.liferay.item.selector.taglib.ReturnType" %><%@
 page import="com.liferay.portal.kernel.dao.search.SearchContainer" %><%@
