+++ modules/frontend/frontend-editors-web/src/META-INF/resources/html/editors/init.jsp
@@ -22,7 +22,7 @@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.document.library.item.selector.web.DLItemSelectorCriterion" %><%@
+<%@ page import="com.liferay.document.library.item.selector.DLItemSelectorCriterion" %><%@
 page import="com.liferay.item.selector.ItemSelector" %><%@
 page import="com.liferay.item.selector.ItemSelectorCriterion" %><%@
 page import="com.liferay.layout.item.selector.web.LayoutItemSelectorCriterion" %><%@
